WIP: Support for variable K offset (reads)
This is a WIP commit adding support for variable K offset reads. The
FieldAccess visitor has been extended to resolve variable K offsets and
dace had to be updated to do proper subset intersection (between ranges
and indices) in MemletDict and MemletSet (iirc these are new in the
stree-to-sdfg branch, so not being super polished makes sense).

I've been testing with `test_variable_offsets` from
`test_code_generation.py` and this test currently fails SDFG validation
with an out of bounds memlet.