fix[next]: Use CustomMapping for caching of offset provider in program call (#2044)
Fix the cache key for the offset provider. We use the
`eve.utils.CustomMapping` class introduced by #2062 to cache the offset
providers in a dictionary and ensure that the key is computed in the
same way at compile time and at runtime. This solution is used in two
places: the cache for the full offset provider, that contains
user-defined and implicit offset providers; and the cache for offset
provider type.

---------

Co-authored-by: Enrique Gonzalez Paredes <enriqueg@cscs.ch>