CI: Limit GT4PY_BUILD_JOBS=2 as workaround for hanging of dace-gpu tests (#2068)
The tests of the dace-gpu CI pipeline are hanging after the latest
maintenance of Santis vCluster. The origin of the problem is not known
yet, but it was observed that limiting the build parallelism of tests
with static args (`GT4PY_BUILD_JOBS=2`) makes the tests pass. Probably
this setting only triggers a different schedule of the tests that hides
the issue. We merge this PR as a workaround to unblock the CI.