feat[dace][next]: Added `MultiStateGlobalSelfCopyElimination2` (#2056)
This PR adds the `MultiStateGlobalSelfCopyElimination2` transformation.
It is very similar to `MultiStateGlobalSelfCopyElimination`, however, it
uses a different way to identify what to remove.
Instead scanning the global data and then identifying it searches for
the transient data.
In fact the two transformation should be merged, however, currently
there is no time in doing that, thus they will exist for the time being
in parallel.