refactor[cartesian]: Always define stencil builder in module generator (#2091)
## Description

Ensure `StencilBuilder` is defined when the `BaseModuleGenerator` is
initialized. The `BaseModuleGenerator` was initially introduced in PR
https://github.com/GridTools/gt4py/pull/376. And as the inline comment
explained, it previous versions, this was not possible. Since then, many
things changed and it looks like nowadays we can always directly
initialize the module generator with a `StencilBuilder` object. This
simplifies the module generator as well as calling code.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
  Existing tests have been updated as needed.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>