fix[next]: Fix temporary allocation in gtfn (#2092)
In gtfn temporary allocation the size was wrongly computed when the
domain of the temporary started in the negative. Now we shift the
temporary such that the origin matches the domain where the temporary is
used (no overallocation).

The fix is in `gtfn/codegen.py`. The rest is infrastructure to support
explicitly creating a temporary at itir level.

---------

Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>