feat[core]: lock api (#2109)
Adds `locking` module to gt4py._core which provides a simple file lock
API: `lock(directory)`.

The default lock is switched back to `filelock` as
- Lustre should support flock
- flufl.lock didn't seem to work nicely for our use-cases (e.g.
performance not perfect and link_count errors)

This PR already allows to customize the locker by overwriting
`locking._lock` but it's not currently publicly exposed.