More improvements.
I am not fully sure why we use the simplify, because they are expensive, but I do not know how much we need them.
Also aovided a `try_initialize()` in parallel case.

The thing is increadably ugly and I think we have to do the refactoring/splitting of the map fusion before we continue in that thing.