fix[dace][next]: No `ScalarToSymbolPromotion` in `gt_auto_optimizer()` (#2132)
In [PR#2085](https://github.com/GridTools/gt4py/pull/2085)
`ScalarToSymbolPropagation` was introduced in the initial simplification
stage to solve a problem with a symbol that had to be created by means
of an InterstateEdge.
However, due to some changes
([PR#2122](https://github.com/GridTools/gt4py/pull/2122) and
[PR#2128](https://github.com/GridTools/gt4py/pull/2128)) in the lowering
this is no longer needed.

Co-authored-by: edopao <edoardo.paone@cscs.ch>