feat[next][dace]: Prevent VerticalMapFusion when there are neighbor reductions (#2124)
Prevents `VerticalMapFusion` when a map writes to an `AccessNode` and
the next map reads from this `AccessNode` using some Neighbor table.
This fusion is restricted because due to the indirect accesses we can't
know which part of the `AccessNode` mentioned before should be read from
the second map. Due to `read-after-write` dependency we have to make
sure that the whole `AccessNode` is written by the first map and then
read by one or more maps.