ci: cleanup cscs cicd-ext pipeline (#2119)
Rewrite the CSCS CI pipeline to make it cleaner and more lightweight.
There is only a single build stage now, and it starts from a heavily
parametrized and versatile dockerfile, which relies on uv to handle all
the python-related package installation.

This is a simplified version of the CSCS CI pipeline developed in #2087
but without the dynamic generation of the child pipeline.

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>