feat[next][dace]: Map symbolic args to nested SDFG (#2122)
**1st and main change**
This PR tries to solve the lowering of symbol assignment in the context
of let statements like `let nlev = vertical_end - 1`.
The let statement is lowered as a nested SDFG. The `nlev` symbol can be
represented on the nested SDFG as either a scalar (the previous
behavior) or a dace symbol. The condition applied by this PR to use the
dace symbolic representation is that all arguments to a scalar
expression are already dace symbols. Thus, the symbolic expression
`vertical_end - 1` can be mapped to the inner symbol `nlev` on the
nested SDFG.

**2nd change**
Suppose a let statement like `let nlev = vertical_end + val - 1` where
`vertical_end` is a dace symbol but `val` is a scalar. In this case, the
lowering would still choose to lower the scalar expression as a tasklet,
because not all arguments are dace symbols, and write the result to a
transient data node. To overcome this limitation, this PR introduces
another change: represent all program inputs as dace symbols. Observe
that this change only applies to the scalar program arguments, that is
the inputs to the top-level SDFG. For scalars on nested SDFGs the same
rules as before still apply.

**3rd change**
The above change enables some cleanup in the check of symbolic
arguments. Since all program scalar arguments are now represented as
dace symbols, we do not need to scan _on the top-level SDFG_ which
symbols are used inside domain expressions. Therefore, this PR moves the
collection of domain symbols to the setup of nested SDFG context.

Note that this PR creates a new module `gtir_lower_types`. In a separate
refactoring PR existing data types should be moved here.