build: update package extras and dependency groups (#2131)
Currently, GT4Py package users cannot rely on the `dace` package extras
to work with a standard python package installation tool, because the
required DaCe versions are not yet available on PyPI. This PR transforms
the extras into dependency groups so they are not longer shown to
package users but are still available for developers and advanced users
with direct access to this repo. The PR also updates the developer
documentation accordingly.

Additionally, this PR contains fixes for cscs-ci and nox.

---------

Co-authored-by: edopao <edoardo16@gmail.com>