Fixed wrong handling of `simplify_at_entry` argument in `gt_substitute_compiletime_symbols()`.
The `gt_substitute_compiletime_symbols()` function provides the undocumented argument `simplify_at_entry` which allows to disable the initial simplification, before the actual substitution.
This is needed to overcome [issue#1817 in DaCe](https://github.com/spcl/dace/issues/1817) and an issue when the substitution of an AccessNode is attempted, an issue that should be fixed by [PR#2079](https://github.com/spcl/dace/pull/2079).
This initial simplification is not needed if the function is called in the context of `gt_auto_optimize()`, where that is done as a first step.
However, for some reason it was using a `not` in front of the check.