fix[next][dace]: Ensure non-negative shape of temporaries in concat_where expressions (#2150)
This PR fixes a bug discovered in icon4py stencil tests. The
implementation in main was performing the check on non-negative array
shape only for the result of concat_where expressions. The problem is
that the strange domain with range `start > stop` can be propagated to
the arguments of concat_where, and therefore it can appear in regular
field operators. This is why the new unittest computes `b * 2` as
argument to concat_where.

The fix consists of applying the check in the general utility function,
also used by regular field operators, for extraction of field shape from
domain range.