Updated the fusability check.
In [PR#2057](https://github.com/GridTools/gt4py/pull/2057) A bug was solved in `MapPromoter`.
The issue was related that the promotion and fusing was done in two different steps, thus the SDFG after promotion was invalid.
The solution was to perform the fusing inside the promotion.

However, the solution was not complete because, it was not checked, if the fusion would really remove the intermediate.
This check was not possible at the time and and a surogate was implemented at the time.
But since DaCe now supports the `require_exclusive_intermediates` argument it is possible to do that.

Also this function reworks the check that the second Map is not empty.