feat[cartesian] Debug Backend (#1884)
## Description

This PR implements a plain python backend for debugging and rapid
prototyping of features.

The main difference between the debug backend and the numpy backend is
that this backend is meant to be easier to step through with a debugger
to understand the generated code or to visually inspect the generated
code because model developers are more familiar with with loop based
implementations.

The second big difference is that this backend directly represents the
OIR without lowering to an additional lower level IR to visualize the
resulting OIR a bit more directly.

This backend is not optimized for speed but rather for readabilty.

The PR comes with a couple of changes to the numpy backend because we
could lift/generalize some code to be reused in the numpy & debug
backends.

## Requirements

- [X] All fixes and/or new features come with corresponding tests. 
- [x] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>
Co-authored-by: Roman Cattaneo <romanc@users.noreply.github.com>
Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>