feat[dace][next]: Optimized `gt_inline_nested_sdfg()`. (#2178)
Before we can inline a nested SDFG, we have to perform a cleaning step
(`PruneSymbols` and `PruneConnectors`) on it. Originally this was done
in a two stage way, i.e. first calling `sdfg.apply_transformations_repeated(
[dace_dataflow.PruneSymbols, dace_dataflow.PruneConnectors], ...)` and
then the inline transformation.

In the new design these two stages were unified. Thus we scan the SDFG
recursively and apply apply the cleaning transformation and the inline
transformation manually. Thus only a single pass through the SDFG is
needed.

For this it was necessary to replicate some functionality inside DaCe into GT4Py.
However, a major difference to DaCe is that we first try to use `InlineSDFG`,
which is only able to handle nested SDFG with a single state, before we try
`MultistateInlineSDFG`. The main reason is that the majority of our SDFGs
only have one state and that `InlineSDFG` does not create additional states.