feat[dace][next]: Less Extensive Validation During `gt_auto_optimizer()` (#2177)
The original idea of the `validate` and `validate_all` flags (which are
inherited from DaCe) are to either perform validation at the very end or
extensively during the operation. This means that `validate` must be set
to `False` for every subset the auto optimizer does. This was missing before.

These changes were previously done in the context of [Optimizing the
Optimizer](https://github.com/GridTools/gt4py/pull/2129).