fix[dace][next]: Disabled `InlineSDFG`. (#2188)
In [PR#2178](https://github.com/GridTools/gt4py/pull/2178) the
`gt_inline_nested_sdfg()` function was modified to be more efficient.
It also changed the order in which the inline transformations of DaCe
where applied.
Instead of trying `InlineMultistateSDFG` it changed that such that
`InlineSDFG` was used, which was desirable, because it did not create
additional state that had to be fused later. However, this triggered
[issue#2108](https://github.com/spcl/dace/issues/2108) which lead to the
removals of some writes.
As a temporary solution we no longer use `InlineSDFG` but only the
multistate version.

---------

Co-authored-by: edopao <edoardo16@gmail.com>