refactor[cartesian, dace]: avoid duplicate transient check in `replace_strides()` (#2184)
## Description

We replacing transient strides, we assemble a list of transient arrays
in `replace_transient_strides()`, which we then pass to
`replace_strides()`. Replace strides will (contrary to it's name) check
again and only replace strides of transient arrays. This is unexpected
and duplicate work. We don't use `replace_strides()` from anywhere else
than `replace_transient_strides()`, so it's "safe" to remove `if
array.transient` even if this alters the functionality of that function
because it doesn't matter in the one use-case that we have.

The change includes fully typing `replace_strides()`, which is achieved
by changing the second argument from `get_layout_map` to include the
full `layout_info` object. That `layout_info` object just has a simpler
type. Doing so, `mypy` started complaining about inconsistent types of
`layout_map` functions, which I fixed as part of this PR.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
  Covered by existing tests.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A