fix[next][dace]: Do not crash if SDFG report not found (#2191)
This PR handles the case where the SDFG are precompiled without
performance metrics `(config.COLLECT_METRICS_LEVEL == metrics.DISABLED)`
but `config.COLLECT_METRICS_LEVEL` is set on another run, when the
stencils are loaded from the binary library. The main baseline was
crashing because the SDFG would return `None` as profiling report.