refactor[cartesian]: invalid frontend/backend raise `ValueError`
This PR changes the error for an invalid frontend/backend choice from
`NotImplementedError` to `ValueError`. While it could happen that the
requested frontend/backend is not yet implemented, the more likely
scenario is that users misspelled an existing frontend/backend name. Imo
this should thus be a value error, in accordance with how the python
docs [1] describe it.

From a practical point of view, `NotImplementedError` is commonly used
in tests with `xfail(raises=NotImplementedError)`, which could - if the
backend is misspelled - hide problems.

[1]: https://docs.python.org/3/library/exceptions.html#ValueError