Unrelated: ValueError for invalid backend choice
Change error for invalid backend choice from NotIpmlementdError to
ValueError. While it could happen that the requested backend is not yet
implemented, the more likely scenario is that users mispelled the
backend name. Imo this should thus return a ValueError.

Found because the (not yet in this PR) debug backend wasn't caught by
`xfail(raises=NotImplementedError)`.