feat[dace][next]: Introduce different Skip Categorieds During Optimization (#2180)
Before all parts of simplify where run every time.
This PR introduces a more fine grained behaviour, by using the `skip`
mechanism, that is already there.
For this three different lists were defined:
- For the initial simplification.
- For phase one, in which the top level dataflow is optimized.
- For phase two, in which the dataflow inside maps is optimized.

The main objective is to improve the performance of the tool chain by
deactivating parts of simplify that are not needed.