ci: add MacOS to daily ci (#2153)
Summary of changes:
- Enable daily CI on `macos` image. However, the default stable version
(`macos-14`) has a compiler issue which is fixed in the next version`
macos-15`. The `macos-15` version will become the default stable version
very soon (in August 2025).
- Minimum version of `nanobind` bumped to `1.9.0` in order to include
fix for a linker issue on macos.
- Skipping check in a dace unittest in gt4py-next because it fails
randomly. The failure becomes consistent after fixing the reduce
initialization in the SDFG. This issue will be addressed in a separate
PR.
- Skipping `test_dace_parsing` in gt4py-cartesian because it fails in
CI.

---------

Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>
Co-authored-by: Ioannis Magkanaris <ioannis.magkanaris@cscs.ch>