Slighty changed how the translator identifies if we are on a GPU.
Before the idientification was performed based on `CUPY_DEVICE_TYPE` and `device_type`.
Since `CUPY_DEVICE_TYPE` depends on if there is a GPU and from which manufactor it is impossible to "force" a particular version, which is usefull in tests.
Thus the tests was changed such that the test was based on CPU.