This should be the solution.
It seems to be important that the compiler is also correctly initialized.
I noticed if I added a breakpoint into `__init__()` of the compiler that I never hit it with the GPU.
After some starring I was finding that.

I have no idea why it only affects sometimes, I guess that it has to do something with the order.