fix[dace][next]: "Fixed" Asynchronous SDFG (#2201)
This PR is related to an earlier change, introduced by
[PR#2189](https://github.com/GridTools/gt4py/pull/2189), which changed
the configuration such that instead of one (random) stream the default
stream of the GPU was used. An unintended side effect was that
asynchronous execution, i.e. returning from `CompiledSDFG.fast_call()`
once all work has been _scheduled_, i.e. not wait until the work has been
completed, which is an [intended](https://github.com/GridTools/gt4py/pull/2035)
behaviour, is no longer an opt-in feature but has become an "always on" feature.

This PR re-enables "synchronous" execution, by adding a final state to the SDFG
that contains a Tasklet which calls the synchronization primitive of the GPU
runtime. This state is added by default, if GPU execution is enabled, to preserve
the previous behaviour.

Note that the changes this PR introduces are potentially dangerous as
they might lead to race conditions. This happens if the result of a GPU
Map is, in any way, not used by a Map. For example, if a GPU Map
computes data, which is then copied back to CPU, where it is used,
either in a Tasklet or on an interstate edge. Because the async
`mempcy()` function is used and there is no synchronization, the
copy might not have concluded before the CPU data is used.
However, the expected impacts are minimal, as in the GT4Py setting this
should not happen as scalar computation are usually done on inputs of
Maps and we do not use interstate edges.

Due to the changes described above some of the DaCe related aspects of
the workflow have been changed. This in turn lead to the discovery of
several bugs, in GT4Py and DaCe, with further consequences. The most
severe one is [DaCe issue#2125](https://github.com/spcl/dace/issues/2125).
The issue is that `dace.Config` is a global object that is shared among all
threads (which is an issue for the tests).