fix[cartesian]: gtscript ValueInliner np.bool_ support
`ValueInliner` in `gtscript` was missing support for `np.bool_` types.
Stencils with externals of type `np.bool_` would end up with an error
saying "Failed to inline {external}.". Since `np.bool_` is part of
gtscript's `_VALID_DATA_TYPES`, we should support externals of
type `np.bool_`.

This PR adds support as well as a test to cover regressions in the
future.