Merge schedule tree roundtrip work into `milestone2` integration branch (#2)
* [unrelated] remove unused function
* Deduplicate k-axis iteration symbols

Deduplicating k-axis iteration symbols because DaCe (v1) uses search &
replace to rename variables when building an stree from a given sdfg.
Without this change we get into situations that have variable shadowing
and then the search & replace algo is too simplistic to handle the
deduplication correctly, e.g. when replacing `__k` in the following

```
int __k;

for(__k = 0; __k < 10; __k += 1) {
	// use __k ...
}

for(int __k = 0; __k < 10; __k += 1) {
	// __k accesses here were replaced leading to issues
}

```

Instead we now do the work upfront and build a different k-axis iteration
symbol into the sdfg that is then transformed to stree in orchestration.

* Update dace to romanc/stree-roundtrip (GT4Py fork)