feat[cartesian]: Support for `round()` / ` round_away_from_zero()` in gtscript (#2198)
# Description

This PR adds support `round()` and `round_away_from_zero()` in
`gtscript`. This PR got a bit of traction because there are basically
two predominant ways to round ties (e.g. when rounding 1.5 to an integer
it's as close to 1.0 as it is to 2.0). The default in python, Kotlin,
Julia, and C# is to split ties by rounding to the nearest even value,
e.g.

- `round(-0.5) == 0.0`
- `round(0.5) == 0.0`
- `round(1.5) == 2.0`
- `round(2.5) == 2.0`

This is also what the IEEE754 floating point standard defines. In
contrast, programming languages like Go, Rust, Ruby, C++, and Fortran
default to round away from zero, e.g.

- `round(-0.5) == -1.0`
- `round(0.5) == 0.0`
- `round(1.5) == 2.0`
- `round(2.5) == 3.0`

From a DSL point of view, it is paramount to implement the same behavior
in all backends, regardless of the programming language of the backend.
We choose to implement `round()` with the tie breaking as suggested by
the standard. In addition, we implement a function
`round_away_from_zero()` which breaks ties by rounding away from zero.
This allows, for example, to truthfully port FORTRAN-based code. The
full discussion and evaluation of other alternatives can be found in the
ADR.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [x] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.

---------

Co-authored-by: Florian Deconinck <deconinck.florian@gmail.com>
Co-authored-by: Roman Cattaneo <>
Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>
Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>