ci: switch to macos-latest in daily-CI (#2239)
This PR addresses a TODO introduced in #2153. The previous PR fixed the
macOS image version to `macOS-15`, in order to address a compiler issue
in Github daily-CI, since `macos-latest` was pointing to `macos-14` at
that time. We can now switch to `macos-latest` since the default image
version has been upgraded.