refactor[cartesian]: delete unused tile symbol function
The DaCe backends currently don't support tiling, so this function is
unused. Even when we add tiling back, we won't do it at the Tree IR
level, but later using existing transformations on the resulting SDFG.
This method will therefore most likely never be used and can safely be
deleted.