fix[next][dace]: follow new canonical field domain order (#2240)
In [PR#1847](github.com/GridTools/gt4py/pull/1847) the convention was
established that the local dimension is always at the end.
However, in [PR#1922](https://github.com/GridTools/gt4py/pull/1847) this
convention was changed such that the local dimension is between the
horizontal and vertical.
However, this convention change was not propagated to the DaCe backend,
which continued to work, since the case, where a field has a horizontal,
vertical and a local was never hit.

This PR introduces this new canonical order into the DaCe backend.
Credits to @philip-paul-mueller, see his original PR #2217.

Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>