fix[cartesian]: actually use cached (to disk) SDFGs (#2246)
## Description

The dace backends cache generated SDFGs in memory and save them to disk
(for later re-use). In-memory caching works as expected. The SDFGs
cached to disk were never read and always re-calculated. This PR changes
the caching logic to

1. check for an existing SDFG in memory
2. check for an existing SDFG in the cache folder on disk
3. calculate (and cache) the SDFG

This saved us a lot of time (e.g. in the benchmark scripts) where we
have all the SFDGs already pre-calculated after a first run.

Note: This is part of untangling the
[`milestone2`](https://github.com/GridTools/gt4py/pull/2202) branch.

## Requirements

- [ ] All fixes and/or new features come with corresponding tests.
  N/A
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A