fix[cartesian]: memlets of data dims in jit workflow (#2247)
## Description

Memlets of data_dims in jit workflows were off by one. The memlet was
configured to include one too many because ranges in dace are inclusive
on both ends (in contrast to many half-open range definitions that are
inclusive at the start and exclusive at the end).

Note: This is part of untangling the
[`milestone2`](https://github.com/GridTools/gt4py/pull/2202) branch.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
  New test added at the NDSL-level where we use the jit workflow
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A